refactor: split commands.go into separate modules by responsibility
- generate_certs.go: GenerateCerts function for CLI cert generation
- run_uncors.go: RunUncors entry point that routes to interactive/non-interactive
- run_ineractive.go: runIneractive for TUI mode (BubbleTea-based)
- run_non_ineractive.go: runNonIneractive for headless mode with config watching

Also:
- Remove --debug CLI flag; logging now controlled solely via UNCORS_LOGGING env var
- Rename parameters for consistency: uncorsConfig→cfg, configPath→cfgPath
- Add context.Context to runIneractive for proper lifecycle management
